# internal/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Tobby REST API
  # The API major version. The binary's build version is shown on /about
  # and in the tobby_build_info metric — this document is static.
  version: v1
  description: >
    Versioned REST API of a Tobby instance, in strict parity with the web
    UI (FR-061): same parameters, same error taxonomy. Payloads carry raw
    bytes and RFC 3339 timestamps exclusively; localization is a UI
    concern. Errors are RFC 9457 problem documents whose extension members
    (code, probable_cause, action, correlation_id) come from the embedded
    error taxonomy; the "type" member is the in-instance troubleshooting
    anchor /help#<code>, resolvable offline.
  license:
    name: GPL-3.0-only
    identifier: GPL-3.0-only
servers:
  - url: /
    description: The instance itself (single listener, ADR-0015).
security:
  - basicAuth: []
  - bearerAuth: []
paths:
...
  /api/v1/content:
    get:
      summary: >-
        List the repositories of the local store, grouped and paginated.
        Mirror of the /content screen. Role viewer.
      parameters:
        - name: q
          in: query
          required: false
          description: Substring filter on the full relocated path.
          schema:
            type: string
        - name: kind
          in: query
          required: false
          description: Keep only the listed kinds (repeatable).
          schema:
            type: array
            items:
              $ref: "#/components/schemas/Kind"
        - name: prefix
          in: query
          required: false
          description: Keep repositories under an exact path prefix.
          schema:
            type: string
        - name: page
          in: query
          required: false
          description: 1-based page number.
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: One page of repositories.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentList"
        default:
          $ref: "#/components/responses/Problem"
  /api/v1/content/{repo}:
    get:
      summary: >-
        Repository detail with its tags. Mirror of the /content/{repo}
        screen. Role viewer.
      parameters:
        - $ref: "#/components/parameters/Repo"
      responses:
        "200":
          description: The repository and its tags.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentRepo"
        default:
          $ref: "#/components/responses/Problem"
...
  /api/v1/accounts:
    get:
      summary: >-
        List the local accounts — never password hashes. Accounts are
        created on the host with the tobby user CLI (R-01). Mirror of the
        /admin/accounts screen. Role admin.
      responses:
        "200":
          description: The accounts.
          content:
            application/json:
              schema:
                type: object
                properties:
                  accounts:
                    type: array
                    items:
                      $ref: "#/components/schemas/Account"
        default:
          $ref: "#/components/responses/Problem"
  /api/v1/tokens:
    get:
      summary: >-
        List the static API tokens, revoked included — never secrets.
        Role admin.
      responses:
        "200":
          description: The tokens.
          content:
            application/json:
              schema:
                type: object
                properties:
                  tokens:
                    type: array
                    items:
                      $ref: "#/components/schemas/Token"
        default:
          $ref: "#/components/responses/Problem"
    post:
      summary: >-
        Create a static API token (FR-072). The clear secret is returned
        exactly once, in this response; only its SHA-256 is stored.
        Role admin.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, role]
              properties:
                name:
                  type: string
                  description: Unique name among the active tokens.
                role:
                  $ref: "#/components/schemas/Role"
      responses:
        "201":
          description: The created token with its one-time secret.
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                  role:
                    $ref: "#/components/schemas/Role"
                  secret:
                    type: string
                    description: >-
                      The clear token secret ("tby_…") — shown exactly
                      once, never re-displayable.
        default:
          $ref: "#/components/responses/Problem"
  /api/v1/tokens/{name}/revoke:
    post:
      summary: >-
        Revoke a token immediately and permanently (FR-072). Role admin.
      parameters:
        - name: name
          in: path
          required: true
          description: The token name.
          schema:
            type: string
      responses:
        "200":
          description: The revoked token.
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    $ref: "#/components/schemas/Token"
        default:
          $ref: "#/components/responses/Problem"
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: >-
        Account name and password, or token name and token secret — the
        same accounts as the web UI and the embedded registry (FR-076).
    bearerAuth:
      type: http
      scheme: bearer
      description: A static API token secret ("tby_…", FR-072).
  parameters:
    Repo:
      name: repo
      in: path
      required: true
      description: >-
        The relocated repository path; may contain slashes
        (e.g. docker.io/library/redis).
      schema:
        type: string
...
  responses:
    Problem:
      description: >-
        Any error, as an RFC 9457 problem document carrying the taxonomy
        entry (application/problem+json) with its entry's HTTP status.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
    Problem:
      type: object
      description: >-
        RFC 9457 problem document. The extension members come from the
        single error taxonomy; "type" is the in-instance troubleshooting
        anchor, resolvable offline.
      properties:
        type:
          type: string
          description: In-instance troubleshooting anchor, "/help#<code>".
        title:
          type: string
          description: What happened, in the negotiated language.
        status:
          type: integer
          description: The HTTP status of the response.
        code:
          type: string
          description: The stable taxonomy code ("TBY-REG-003").
        probable_cause:
          type: string
          description: The probable cause, in the negotiated language.
        action:
          type: string
          description: The corrective action, in the negotiated language.
        correlation_id:
          type: string
          description: Run or task identifier for log correlation (FR-090).
      required: [type, title, status, code, probable_cause, action]
    Kind:
      type: string
      description: The five frozen ingredient kinds (never translated).
      enum: [ContainerImage, HelmChart, OCIArtifact, FileSet, Recipe]
    Role:
      type: string
      description: The closed role ladder (ADR-0009).
      enum: [viewer, operator, admin]
...
    ContentList:
      type: object
      properties:
        repositories:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              kind:
                $ref: "#/components/schemas/Kind"
              tags:
                type: integer
              sizeBytes:
                type: integer
                description: Logical size in bytes (deduplicated per repository).
        page:
          type: integer
        totalPages:
          type: integer
        total:
          type: integer
    ContentRepo:
      type: object
      properties:
        name:
          type: string
        kind:
          $ref: "#/components/schemas/Kind"
        tags:
          type: array
          items:
            type: object
            properties:
              tag:
                type: string
              digest:
                type: string
              sizeBytes:
                type: integer
              platforms:
                type: integer
...
    Account:
      type: object
      properties:
        name:
          type: string
        role:
          $ref: "#/components/schemas/Role"
        created:
          type: string
          format: date-time
        last_login:
          type: string
          format: date-time
    Token:
      type: object
      properties:
        name:
          type: string
        role:
          $ref: "#/components/schemas/Role"
        created:
          type: string
          format: date-time
        revoked:
          type: boolean